Update BetControls.vue
First try on solving https://github.com/BetKingIO/SampleApp/issues/6
Added Balance on mapGetters since  this.Balance is required by bet method.
Remove disable on BetProfit, added updateBetAmount function
On every keyup updateBetAmount function checks if value on BetProfit is 'NaN' or < 0, if it is, set BetAmount  to 0. 
If BetProfit is > 0,  we calculate the BetAmount by ( BetProfit / (Payout -1) ).  If BetAmount is greater than the Balance, set BetAmount to its Balance resulting to a max bet. 
And lastly we call updateProfit.